+++ src/modules/grids/GridsView.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from "react";
 import {
   StyleSheet,
   View,
@@ -8,11 +8,11 @@ import {
   TouchableOpacity,
   Image,
   Dimensions,
-} from 'react-native';
-import LinearGradient from 'react-native-linear-gradient';
-import { colors, fonts } from '../../styles';
+} from "react-native";
+import LinearGradient from "react-native-linear-gradient";
+import { colors, fonts } from "../../styles";
 
-import { RadioGroup, GridRow } from '../../components';
+import { RadioGroup, GridRow } from "../../components";
 
 export default class GridsScreen extends React.Component {
   _getRenderItemFunction = () =>
@@ -20,14 +20,14 @@ export default class GridsScreen extends React.Component {
       this.props.tabIndex
     ];
 
-  _openArticle = article => {
-    this.props.navigation.navigate('Article', {
+  _openArticle = (article) => {
+    this.props.navigation.navigate("Article", {
       article,
     });
   };
 
-  renderRowOne = rowData => {
-    const cellViews = rowData.item.map(item => (
+  renderRowOne = (rowData) => {
+    const cellViews = rowData.item.map((item) => (
       <TouchableOpacity key={item.id} onPress={() => this._openArticle(item)}>
         <View style={styles.itemOneContainer}>
           <View style={styles.itemOneImageContainer}>
@@ -100,7 +100,7 @@ export default class GridsScreen extends React.Component {
               <View
                 style={[
                   styles.badge,
-                  item.badge === 'NEW' && { backgroundColor: colors.labelOne },
+                  item.badge === "NEW" && { backgroundColor: colors.labelOne },
                 ]}
               >
                 <Text
@@ -129,14 +129,14 @@ export default class GridsScreen extends React.Component {
       <View style={styles.container}>
         <View style={{ height: 50 }}>
           <RadioGroup
-            xtedIndex={this.props.tabIndex}
+            selectedIndex={this.props.tabIndex}
             items={this.props.tabs}
             onChange={this.props.setTabIndex}
             underline
           />
         </View>
         <FlatList
-          keyExtractor={item =>
+          keyExtractor={(item) =>
             item.id
               ? `${this.props.tabIndex}-${item.id}`
               : `${item[0] && item[0].id}`
@@ -156,20 +156,20 @@ const styles = StyleSheet.create({
     backgroundColor: colors.white,
   },
   tabsContainer: {
-    alignSelf: 'stretch',
+    alignSelf: "stretch",
     marginTop: 30,
   },
   itemOneContainer: {
     flex: 1,
-    width: Dimensions.get('window').width / 2 - 40,
+    width: Dimensions.get("window").width / 2 - 40,
   },
   itemOneImageContainer: {
     borderRadius: 3,
-    overflow: 'hidden',
+    overflow: "hidden",
   },
   itemOneImage: {
     height: 200,
-    width: Dimensions.get('window').width / 2 - 40,
+    width: Dimensions.get("window").width / 2 - 40,
   },
   itemOneTitle: {
     fontFamily: fonts.primaryRegular,
@@ -178,7 +178,7 @@ const styles = StyleSheet.create({
   itemOneSubTitle: {
     fontFamily: fonts.primaryRegular,
     fontSize: 13,
-    color: '#B2B2B2',
+    color: "#B2B2B2",
     marginVertical: 3,
   },
   itemOnePrice: {
@@ -186,8 +186,8 @@ const styles = StyleSheet.create({
     fontSize: 15,
   },
   itemOneRow: {
-    flexDirection: 'row',
-    justifyContent: 'space-around',
+    flexDirection: "row",
+    justifyContent: "space-around",
     marginTop: 10,
   },
   itemOneContent: {
@@ -196,13 +196,13 @@ const styles = StyleSheet.create({
   },
   itemTwoContainer: {
     paddingBottom: 10,
-    backgroundColor: 'white',
+    backgroundColor: "white",
     marginVertical: 5,
   },
   itemTwoContent: {
     padding: 20,
-    position: 'relative',
-    marginHorizontal: Platform.OS === 'ios' ? -15 : 0,
+    position: "relative",
+    marginHorizontal: Platform.OS === "ios" ? -15 : 0,
     height: 150,
   },
   itemTwoTitle: {
@@ -222,14 +222,14 @@ const styles = StyleSheet.create({
     fontSize: 20,
   },
   itemTwoImage: {
-    position: 'absolute',
+    position: "absolute",
     top: 0,
     left: 0,
     right: 0,
     bottom: 0,
   },
   itemTwoOverlay: {
-    position: 'absolute',
+    position: "absolute",
     left: 0,
     right: 0,
     top: 0,
@@ -237,10 +237,10 @@ const styles = StyleSheet.create({
     opacity: 0.5,
   },
   itemThreeContainer: {
-    backgroundColor: 'white',
+    backgroundColor: "white",
   },
   itemThreeSubContainer: {
-    flexDirection: 'row',
+    flexDirection: "row",
     paddingVertical: 10,
   },
   itemThreeImage: {
@@ -250,38 +250,38 @@ const styles = StyleSheet.create({
   itemThreeContent: {
     flex: 1,
     paddingLeft: 15,
-    justifyContent: 'space-between',
+    justifyContent: "space-between",
   },
   itemThreeBrand: {
     fontFamily: fonts.primaryRegular,
     fontSize: 14,
-    color: '#617ae1',
+    color: "#617ae1",
   },
   itemThreeTitle: {
     fontFamily: fonts.primaryBold,
     fontSize: 16,
-    color: '#5F5F5F',
+    color: "#5F5F5F",
   },
   itemThreeSubtitle: {
     fontFamily: fonts.primaryRegular,
     fontSize: 12,
-    color: '#a4a4a4',
+    color: "#a4a4a4",
   },
   itemThreeMetaContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
   },
   itemThreePrice: {
     fontFamily: fonts.primaryRegular,
     fontSize: 15,
-    color: '#5f5f5f',
-    textAlign: 'right',
+    color: "#5f5f5f",
+    textAlign: "right",
   },
   itemThreeHr: {
     flex: 1,
     height: 1,
-    backgroundColor: '#e3e3e3',
+    backgroundColor: "#e3e3e3",
     marginRight: -15,
   },
   badge: {