+++ src/modules/navigation/MainTabNavigator.js
@@ -1,9 +1,9 @@
-import * as React from 'react';
-import { Text, View, Image, StyleSheet } from 'react-native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { colors } from '../../styles';
+import * as React from "react";
+import { Text, View, Image, StyleSheet } from "react-native";
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import { colors } from "../../styles";
 
-import tabNavigationData from './tabNavigationData';
+import tabNavigationData from "./tabNavigationData";
 
 const Tab = createBottomTabNavigator();
 
@@ -11,33 +11,46 @@ export default function BottomTabs() {
   return (
     <Tab.Navigator>
       {tabNavigationData.map((item, idx) => (
-        <Tab.Screen 
-          key={`tab_item${idx+1}`}
+        <Tab.Screen
+          key={`tab_item${idx + 1}`}
           name={item.name}
           component={item.component}
           options={{
-          tabBarIcon: ({ focused }) => (
-            <View style={styles.tabBarItemContainer}>
-              <Image
-                resizeMode="contain"
-                source={item.icon}
-                style={[styles.tabBarIcon, focused && styles.tabBarIconFocused]}
-              />
-            </View>
-          ),
-          tabBarLabel: ({ focused }) => <Text style={{ fontSize: 12, color: focused ? colors.primary : colors.gray }}>{item.name}</Text>
-        }} 
-        />        
+            headerShown: false,
+            tabBarIcon: ({ focused }) => (
+              <View style={styles.tabBarItemContainer}>
+                <Image
+                  resizeMode="contain"
+                  source={item.icon}
+                  style={[
+                    styles.tabBarIcon,
+                    focused && styles.tabBarIconFocused,
+                  ]}
+                />
+              </View>
+            ),
+            tabBarLabel: ({ focused }) => (
+              <Text
+                style={{
+                  fontSize: 12,
+                  color: focused ? colors.primary : colors.gray,
+                }}
+              >
+                {item.name}
+              </Text>
+            ),
+          }}
+        />
       ))}
     </Tab.Navigator>
   );
-};
+}
 
 const styles = StyleSheet.create({
   tabBarItemContainer: {
     flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
+    alignItems: "center",
+    justifyContent: "center",
     borderBottomWidth: 2,
     borderBottomColor: colors.white,
     paddingHorizontal: 10,